fix raii in stbi img handle

+++ src/widget/widget.h
@@ -149,7 +149,7 @@ public:
                                    label_format_Implot_axis.data());
             ImPlot::PlotBars(label.c_str(), data.data(), data.size());
             // Utils::Log::logVec(data);
-            for (int i = 0; i < pos.size(); i++) {
+            for (uint64_t i = 0; i < pos.size(); i++) {
                 ImPlot::Annotation(pos[i], data[i], ImVec4(0, 0, 0, 0), ImVec2(0, -10), true,
                                    "%.1f", data[i]);
             }
